requirements.txt
# VLR_AI Trading System - Complete Production Dependencies
# Updated and consolidated for REAL trading operations

# Core Trading APIs
dhanhq>=2.0.2
kiteconnect>=5.0.1

# Data Processing & Analysis
pandas>=1.5.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.2.0

# Web and HTTP
requests>=2.28.0
aiohttp>=3.8.0
urllib3>=1.26.0
httpx>=0.24.0
websockets>=11.0.0

# Configuration and Environment
python-dotenv>=0.21.0
pyyaml>=6.0

# Market Data & Financial APIs
yfinance>=0.2.18
alpha-vantage>=2.3.1

# Scheduling and Automation
schedule>=1.2.0
apscheduler>=3.10.0

# Excel/Sheets Integration
openpyxl>=3.1.0
xlsxwriter>=3.1.0
gspread>=5.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0

# Authentication and Security
pyotp>=2.6.0
selenium>=4.0.0
cryptography>=40.0.0
bcrypt>=4.0.0

# System Monitoring & Performance
psutil>=5.9.0
memory-profiler>=0.60.0

# Date and Time
python-dateutil>=2.8.0
pytz>=2022.1

# Dashboard & Visualization
plotly>=5.14.0
dash>=2.10.0
dash-bootstrap-components>=1.4.0
flask>=2.3.0

# Caching & Database
redis>=4.5.0
sqlalchemy>=2.0.0

# Async & Concurrency
asyncpg>=0.28.0
aiofiles>=23.0.0

# Technical Analysis
ta-lib>=0.4.25

# Logging & Monitoring
loguru>=0.7.0
structlog>=23.0.0

# Email & Notifications
email-validator>=2.0.0

# Testing & Development
pytest>=7.3.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0

# Data Validation
pydantic>=1.10.0
marshmallow>=3.19.0

# Mathematical & Statistical
statsmodels>=0.14.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Parallel Processing
joblib>=1.2.0

# Error Handling & Recovery
tenacity>=8.2.0
backoff>=2.2.0

# Production Dependencies
gunicorn>=20.1.0
supervisor>=4.2.0

# Additional packages for enhanced functionality
fastapi>=0.100.0
uvicorn>=0.23.0
celery>=5.3.0
flower>=2.0.0
